requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
Werkzeug==2.0.3
email-validator==1.1.3
Pillow==10.2.0
gunicorn==20.1.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
# Optional: Uncomment if using PostgreSQL
# psycopg2-binary==2.9.1